+++ activestorage/db/update_migrate/20211119233751_remove_not_null_on_active_storage_blobs_checksum.rb
@@ -1,0 +1,5 @@
+class RemoveNotNullOnActiveStorageBlobsChecksum < ActiveRecord::Migration[6.0]
+  def change
+    change_column_null(:active_storage_blobs, :checksum, true)
+  end
+end