+++ activestorage/lib/active_storage.rb
@@ -64,6 +64,7 @@ module ActiveStorage
   mattr_accessor :urls_expire_in
 
   mattr_accessor :routes_prefix, default: "/rails/active_storage"
+  mattr_accessor :key_interpolation_procs, default: {}
   mattr_accessor :draw_routes, default: true
   mattr_accessor :resolve_model_to_route, default: :rails_storage_redirect