+++ activestorage/CHANGELOG.md
@@ -247,3 +247,21 @@
 
 
 Please check [6-1-stable](https://github.com/rails/rails/blob/6-1-stable/activestorage/CHANGELOG.md) for previous changes.
+
+
+*   Add ability to customize the Blob#key per Model and attribute basis. Active Storage will save the Blob's attachment on the specified service at the configured key. You can also _interpolate_ values in it automatically with configurable procs.
+
+    ```ruby
+    # app/models/user.rb
+    has_one_attached :avatar,
+                     key: ':tenant/users/:record_id/avatar/:blob_checksum'
+
+    # with the following configuration
+    config.active_storage.key_interpolation_procs = {
+      tenant:        ->(record, blob) { Apartment::Tenant.current.parameterize },
+      record_id:     ->(record, blob) { record.id },
+      blob_checksum: ->(record, blob) { blob.checksum }
+    }
+    ```
+
+    *František Rokůsek*